algorithm

+++ Algorithm/week6/fibonacci_recursion.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-int fib(int n)
+long long int fib(int n)
 {
     if (n <= 2)
         return 1;
